refactoring benchmarks
